+++ blocks/form/form.js
@@ -1,4 +1,5 @@
 import { readBlockConfig, toCamelCase } from '../../scripts/lib-franklin.js';
+import { uploadFile } from './uploadfile.js';
 
 const formatFns = await (async function imports() {
   try {
@@ -171,22 +172,6 @@ function createHidden(fd) {
   return input;
 }
 
-const uploadFile = async (fileInput, fileUploadUrl) => {
-  // Currently supporting single file upload
-  if(fileInput && fileInput?.files?.length > 0 && fileUploadUrl) {
-    let formData = new FormData();  
-    formData.append("file", fileInput.files[0]);
-    let init = {
-      method: 'POST',
-      body: formData
-    }
-    let response = await fetch(fileUploadUrl, init);
-    let result = await response.text();
-    fileInput.dataset.value = result;
-    return response.ok
-  }
-}
-
 function createFile(fd) {
   const field = createFieldWrapper(fd);
   const fileInput = createInput(fd);
@@ -195,14 +180,14 @@ function createFile(fd) {
 
   field.append(fileInput);
   field.append(status);
-  fileInput.onchange = (async (event) => {
-      const form = fileInput?.closest("form");
-      const submitButton = form.querySelector('button[type="submit"]');
-      submitButton ? submitButton.disabled = true : null;
-      status.textContent = "Uploading..." // TODO - localization
-      const resp = await uploadFile(fileInput, form.dataset.fileuploadurl);
-      submitButton ? submitButton.disabled = false : null;
-      status.textContent = resp ? "Uploaded Successfully" : "Upload failed";
+  fileInput.addEventListener('change', async (event) => {
+    const form = fileInput?.closest("form");
+    const submitButton = form.querySelector('button[type="submit"]');
+    submitButton ? submitButton.disabled = true : null;
+    status.textContent = "Uploading..." // TODO - localization
+    const resp = await uploadFile(fileInput, form.dataset.fileuploadurl);
+    submitButton ? submitButton.disabled = false : null;
+    status.textContent = resp ? "Uploaded Successfully" : "Upload failed";
   })
   return field;
 }
